updates prepping for new release

+++ db/migrate/20130813175358_create_inventory_checks.rb
@@ -7,15 +7,15 @@ class CreateInventoryChecks < ActiveRecord::Migration
     	t.string :forty_mg_lot_number
     	t.integer :forty_mg_box_inventory
     	t.integer :forty_mg_bottle_inventory
-    	t.float :forty_mg_open_bottle_inventory
-    	t.float :forty_mg_tablet_count
-    	t.float :forty_mg_total_inventory
+    	t.integer :forty_mg_open_bottle_inventory
+    	t.integer :forty_mg_tablet_count
+    	t.integer :forty_mg_total_inventory
     	t.string :five_mg_lot_number
     	t.integer :five_mg_box_inventory
     	t.integer :five_mg_bottle_inventory
-    	t.float :five_mg_open_bottle_inventory
-    	t.float :five_mg_tablet_count
-    	t.float :five_mg_total_inventory
+    	t.integer :five_mg_open_bottle_inventory
+    	t.integer :five_mg_tablet_count
+    	t.integer :five_mg_total_inventory
     	t.integer :total_inventory_count
 	
       t.timestamps
